requirements.txt
anthropic==0.42.0
rich==13.9.4
python-dotenv==1.0.1